# docs/swagger.yaml
definitions:
  dto.CommentRequest:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    type: object
  dto.CommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      user_id:
        type: integer
    type: object
  dto.DeleteCommentResponse:
    properties:
      message:
        type: string
    type: object
  dto.DeletePhotoResponse:
    properties:
      message:
        type: string
    type: object
  dto.DeleteSocialMediaResponse:
    properties:
      message:
        type: string
    type: object
  dto.DeleteUserResponse:
    properties:
      message:
        type: string
    type: object
  dto.EditCommentRequest:
    properties:
      message:
        type: string
    type: object
  dto.EditCommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.EmbeddedPhotoResponse:
    properties:
      caption:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
  dto.EmbeddedUser:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.EmbeddedUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.GetCommentResponse:
    properties:
      Photo:
        $ref: '#/definitions/dto.EmbeddedPhotoResponse'
      User:
        $ref: '#/definitions/dto.EmbeddedUserResponse'
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.GetPhotoResponse:
    properties:
      User:
        $ref: '#/definitions/dto.EmbeddedUserResponse'
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.GetSocialMediaResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.EmbeddedUser'
      user_id:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  dto.PhotoRequest:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    type: object
  dto.PhotoResponse:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
  dto.PhotoUpdateResponse:
    properties:
      caption:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.RegisterRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.RegisterResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.SocialMediaRequest:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    type: object
  dto.SocialMediaResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      user_id:
        type: integer
    type: object
  dto.UpdateUserDataRequest:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  dto.UpdateUserDataResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /comments:
    get:
      consumes:
      - application/json
      description: Get All comment by identify user from token
      parameters:
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GetCommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create new comment
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.CommentRequest'
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      tags:
      - Comment
  /comments/{commentID}:
    delete:
      consumes:
      - application/json
      description: Delete comment by identify user from token
      parameters:
      - default: "2"
        description: Comment ID
        in: path
        name: commentID
        required: true
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteCommentResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update comment by identify user from token
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.EditCommentRequest'
      - default: "1"
        description: Comment ID
        in: path
        name: commentID
        required: true
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditCommentResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      tags:
      - Comment
  /photos:
    get:
      consumes:
      - application/json
      description: Get All Photos based on token user
      parameters:
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYyODU5MzYsImlkIjoyfQ.ahlQ3Icb2h72Aam9jWgZFM-TultwjKzymM7DGKHVVKU
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GetPhotoResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            type: object
      tags:
      - Photos
    post:
      consumes:
      - application/json
      description: Post Photo based on token user
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.PhotoRequest'
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYyODU5MzYsImlkIjoyfQ.ahlQ3Icb2h72Aam9jWgZFM-TultwjKzymM7DGKHVVKU
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PhotoResponse'
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      tags:
      - Photos
  /photos/{photoID}:
    delete:
      consumes:
      - application/json
      description: Delete Photo based on token user with paramId of photoId
      parameters:
      - default: "2"
        description: Photo ID
        in: path
        name: photoID
        required: true
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYyODU5MzYsImlkIjoyfQ.ahlQ3Icb2h72Aam9jWgZFM-TultwjKzymM7DGKHVVKU
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeletePhotoResponse'
        "403":
          description: Forbidden
          schema:
            type: object
      tags:
      - Photos
    put:
      consumes:
      - application/json
      description: Update Photo based on token user with paramId of photoId
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.PhotoRequest'
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYyODU5MzYsImlkIjoyfQ.ahlQ3Icb2h72Aam9jWgZFM-TultwjKzymM7DGKHVVKU
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - default: "1"
        description: Photo ID
        in: path
        name: photoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.PhotoUpdateResponse'
        "403":
          description: Forbidden
          schema:
            type: object
      tags:
      - Photos
  /social-medias:
    get:
      consumes:
      - application/json
      description: Add social media
      parameters:
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GetSocialMediaResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: object
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: Add social media
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.SocialMediaRequest'
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SocialMediaResponse'
        "404":
          description: Not Found
          schema:
            type: object
      tags:
      - Social Media
  /social-medias/{socmedId}:
    delete:
      consumes:
      - application/json
      description: Delete social media data
      parameters:
      - default: "2"
        description: Social Media ID
        in: path
        name: socmedId
        required: true
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteSocialMediaResponse'
        "404":
          description: Not Found
          schema:
            type: object
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Edit social media data
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.SocialMediaRequest'
      - default: "1"
        description: Social Media ID
        in: path
        name: socmedId
        required: true
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYzNjc3MjcsImlkIjoyfQ.T5NaUHAkGra7ehNfJ09rCLUdz9uDfjTYSbj1EtSbg5Y
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SocialMediaResponse'
        "404":
          description: Not Found
          schema:
            type: object
      tags:
      - Social Media
  /users:
    delete:
      consumes:
      - application/json
      description: Delete User
      parameters:
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImJAYi5jb20iLCJleHAiOjE2NjYyODYxNTUsImlkIjo1fQ.VWqZ38H-gi87mwL3A75YloIBA27JqucEMZsR2TJtIBU
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Only can update username & email only
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDataRequest'
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFAYS5jb20iLCJleHAiOjE2NjYyODU5MzYsImlkIjoyfQ.ahlQ3Icb2h72Aam9jWgZFM-TultwjKzymM7DGKHVVKU
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateUserDataResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login for existing user
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      summary: Login user
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: User Request
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            type: object
      summary: Register new user
      tags:
      - Users
swagger: "2.0"